P2p chain (#12)
* Add the ability to connect blocks to p2p-chain

* docs: Add more documentation to stream_reader

* p2p-chain: Allow node to send requests to peer

* Download headers from peer

* Initial block download working from wire

* few tweaks on ibd

* Abstract block download into the proper struct

* Add timeout for requests

* Add some extra debugs

* p2p_chain: Add address manager

* rebase and solve some conflicts

* refactor: Move node logic to node.rs and remove unused

* remove unused "protocol"

* implement address_man

This module keeps a list of addresses that our node can use to make
new connections.

* Update confuse log

Both blockchain and electrum says new peer. Now electrum says new
client.

* fix some bugs on ibd

* add feature guard on error.rs

* Use UtreexoBlock intead of block

* Docs: Document block_download module

* Fix bug with read loop

* Start downloading proofs from the network

* finish block download logic using utreexo p2p messages

* Remove unused RPC

* Try to save a few loop iterations

* Change log order and ask for blocks before timeout

* Remove allow(unused)

* Fix linting

* Fix some peer bugs and improve logging

* hardens address_man

* Remove debug print

* Update internal information for address

* Don't log error as info

* Remove allow(unused) from peer

* Remove allow(unused) from udata

* Remove allow(unused) from request

* fix some sync bugs

* Doc: add module-level docs to address_man

* Update visibility of proof_utils

+++ src/blockchain/cli_blockchain/mod.rs
@@ -259,7 +259,7 @@ impl UtreexodBackend {
             self.chainstate.flush()?;
             info!(
                 "Downloading headers...{:02.0}%",
-                ((i * 2_000) as f64) / ((tip as f64) as f64) * 100.00
+                ((i * 2_000) as f64) / (tip as f64) * 100.00
             );
         }
         Ok(())
